refactor(api): migrate get-certification-point-of-contact usecase
in identity-access-management
